Ensure sdist also contains the right pre-release tag
Since we don't have access to git metadata when building from a sdist, we
record this information in a separate file `git_extra_version`, which
is generated when creating the sdist.
