+++ python/equistore-core/MANIFEST.in
@@ -3,4 +3,5 @@ include equistore-core-*.crate
 include pyproject.toml
 include AUTHORS
 include LICENSE
-include version_from_git.py
+
+include git_extra_version
