+++ python/equistore-operations/MANIFEST.in
@@ -1,4 +1,5 @@
 include pyproject.toml
 include AUTHORS
 include LICENSE
-include version_from_git.py
+
+include git_extra_version
