+++ python/equistore-operations/setup.py
@@ -5,6 +5,7 @@ import uuid
 
 from setuptools import setup
 from setuptools.command.bdist_egg import bdist_egg
+from setuptools.command.sdist import sdist
 
 
 ROOT = os.path.realpath(os.path.dirname(__file__))
@@ -25,6 +26,22 @@ class bdist_egg_disabled(bdist_egg):
             + "uninstall equistore -y && pip install dist/equistore-*.whl` "
             + "to install from source."
         )
+
+
+class sdist_git_version(sdist):
+    """
+    Create a sdist with an additional generated file containing the extra
+    version from git.
+    """
+
+    def run(self):
+        with open("git_extra_version", "w") as fd:
+            fd.write(git_extra_version())
+
+        # run original sdist
+        super().run()
+
+        os.unlink("git_extra_version")
 
 
 def git_extra_version():
@@ -85,6 +102,16 @@ def git_extra_version():
 
 
 if __name__ == "__main__":
+    if os.path.exists("git_extra_version"):
+        # we are building from a sdist, without git available, but the git
+        # version was recorded in a git_extra_version file
+        with open("git_extra_version") as fd:
+            extra_version = fd.read()
+    else:
+        extra_version = git_extra_version()
+
+    version = "0.1.0" + extra_version
+
     with open(os.path.join(ROOT, "AUTHORS")) as fd:
         authors = fd.read().splitlines()
 
@@ -92,8 +119,6 @@ if __name__ == "__main__":
         # handle "raw" symlink files (on Windows or from full repo tarball)
         with open(os.path.join(ROOT, authors[0])) as fd:
             authors = fd.read().splitlines()
-
-    version = "0.1.0" + git_extra_version()
 
     install_requires = []
     if os.path.exists(EQUISTORE_CORE):
@@ -113,5 +138,6 @@ if __name__ == "__main__":
         install_requires=install_requires,
         cmdclass={
             "bdist_egg": bdist_egg if "bdist_egg" in sys.argv else bdist_egg_disabled,
+            "sdist": sdist_git_version,
         },
     )
