+++ python/equistore-torch/MANIFEST.in
@@ -1,7 +1,8 @@
 include pyproject.toml
 include AUTHORS
 include LICENSE
-include version_from_git.py
+
+include git_extra_version
 
 include equistore-torch.tar.gz
