--- python/equistore/version_from_git.py
@@ -1,58 +1,0 @@
-import subprocess
-
-
-def git_extra_version():
-    """
-    If git is available, it is used to check if we are installing a development
-    version or a released version (by checking how many commits happened since
-    the last tag).
-    """
-
-    # Add pre-release info the version
-    try:
-        tags_list = subprocess.run(
-            ["git", "tag"],
-            stderr=subprocess.DEVNULL,
-            stdout=subprocess.PIPE,
-            check=True,
-        )
-        tags_list = tags_list.stdout.decode("utf8").strip()
-
-        if tags_list == "":
-            first_commit = subprocess.run(
-                ["git", "rev-list", "--max-parents=0", "HEAD"],
-                stderr=subprocess.DEVNULL,
-                stdout=subprocess.PIPE,
-                check=True,
-            )
-            reference = first_commit.stdout.decode("utf8").strip()
-
-        else:
-            last_tag = subprocess.run(
-                ["git", "describe", "--tags", "--abbrev=0"],
-                stderr=subprocess.DEVNULL,
-                stdout=subprocess.PIPE,
-                check=True,
-            )
-
-            reference = last_tag.stdout.decode("utf8").strip()
-
-    except Exception:
-        reference = ""
-        pass
-
-    try:
-        n_commits_since_tag = subprocess.run(
-            ["git", "rev-list", f"{reference}..HEAD", "--count"],
-            stderr=subprocess.DEVNULL,
-            stdout=subprocess.PIPE,
-            check=True,
-        )
-        n_commits_since_tag = n_commits_since_tag.stdout.decode("utf8").strip()
-
-        if n_commits_since_tag != 0:
-            return ".dev" + n_commits_since_tag
-    except Exception:
-        pass
-
-    return ""
